Revert "Moar testing, Config as its own class, fix webservers state handling!"
