+++ tests/test_plight.py
@@ -1,49 +1,53 @@
 import pytest
+import plight
 
-def test_default_nodestatus(status, states):
-    """
-    There are only two default states defined in the
-    test cases. This looks for them explicitly
-    and asserts if neither is found.
-    """
-    found = False
-    for default in ['enabled', 'master']:
-        if default in states.keys():
-            assert status.get_node_state() == default.upper()
-            found = True
-    assert found
+@pytest.fixture(scope='module')
+def status(request):
+    status = plight.NodeStatus()
+    def reset():
+        status.set_node_enabled()
+    request.addfinalizer(reset)
+    return status
 
-def test_default_nodestatus_details(status, states):
-    """
-    Ensure that the NodeStatus object maintains data to match
-    the provided states
-    """
-    if status._default_state in states.keys():
-        for (state, data) in states.items():
-            for (detail, value) in data.items():
-                assert status.get_state_detail(detail, state) == value
-                if state == 'master' and detail == 'message':
-                    assert status.get_state_detail(detail, state) == 'master node'
+# Test enabling
+def test_default_nodestatus(status):
+    assert status.get_node_state() == 'ENABLED'
 
-def test_change_nodestatus(status, states):
-    """
-    Loop through every state and ensure that it is changed
-    to successfully
-    """
-    for (state, data) in states.items():
-        status.set_node_state(data['command'])
-        assert status.get_node_state() == state.upper()
+def test_reset_state_enabled_from_disabled_nodestatus(status):
+    status.set_node_state('disable')
+    status.set_node_state('enable')
+    assert status.get_node_state() == 'ENABLED'
 
-def test_reset_state_to_default_nodestatus(status, states):
-    """
-    Loop though every state, then reset to the default state
-    and verify it behaved as expected
-    """
-    found = False
-    for (state, data) in states.items():
-        status.set_node_state(data['command'])
-        if status._default_state in ['enabled', 'master']:
-            status.set_node_state(status._default_state)
-            assert status.get_node_state() == status._default_state.upper()
-            found = True
-    assert found
+def test_reset_state_enabled_from_offline_nodestatus(status):
+    status.set_node_state('offline')
+    status.set_node_state('enable')
+    assert status.get_node_state() == 'ENABLED'
+
+# Test disabling
+def test_set_disabled_nodestatus(status):
+    status.set_node_disabled()
+    assert status.get_node_state() == 'DISABLED'
+
+def test_set_state_disabled_nodestatus(status):
+    status.set_node_state('disable')
+    assert status.get_node_state() == 'DISABLED'
+
+def test_reset_state_disabled_from_offline_nodestatus(status):
+    status.set_node_state('offline')
+    status.set_node_state('disable')
+    assert status.get_node_state() == 'DISABLED'
+
+# Test offline
+def test_set_offline_nodestatus(status):
+    status.set_node_offline()
+    assert status.get_node_state() == 'OFFLINE'
+
+def test_set_state_offline_nodestatus(status):
+    status.set_node_state('offline')
+    assert status.get_node_state() == 'OFFLINE'
+
+def test_reset_state_offline_from_disabled_nodestatus(status):
+    status.set_node_state('disable')
+    status.set_node_state('offline')
+    assert status.get_node_state() == 'OFFLINE'
+
